requirements.txt
click==8.0.3
colorama==0.4.4
cycler==0.11.0
fonttools==4.28.5
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
tomli==2.0.0
typing_extensions==4.0.1
